+++ robotframework-ls/src/robotframework_ls/robotframework_ls_impl.py
@@ -498,6 +498,8 @@ class RobotFrameworkLanguageServer(PythonLanguageServer):
                     "refactor",
                     "refactor.extract.local",
                     "assign.toVar",
+                    "surroundWith.tryExcept",
+                    "surroundWith.tryExceptFinally",
                 ],
             },
             "codeLensProvider": {"resolveProvider": True},